requirements.txt
flask>=3.0.0

# Gemini SDK
google-genai>=1.56.0

# Google auth & Sheets
google-auth>=2.45.0
google-auth-oauthlib>=1.2.3
google-api-python-client>=2.120.0

# Env
python-dotenv>=1.0.1
